feat(nextdns): add configurable TTL and cache to nextdns
This brings in two new configuration parameters, `NEXTDNS_MAX_TTL` and
`NEXTDNS_CACHE_SIZE` as optional parameters for configuring the NextDNS
CLI.

`NEXTDNS_MAX_TTL`: the maximum TTL for domains returned by NextDNS in
seconds. This defaults to `15`, but can be set higher particularly if
you've got caching layers elsewhere in your stack.

`NEXTDNS_CACHE_SIZE`: the size of NextDNS cache. Most systems should not
need to change this. It defaults to `10MB` - note that this requires a
unit.

Both are set as environment variables that can be imported by the
container.

DCO-1.1-Signed-off-by: Patrick Wagstrom <patrick@wagstrom.net>